// angular-admin/src/app/dto/vehiculo-dto.ts
export interface VehiculoDto {
    combustion:         string;
    modelo:             string;
    transmision:        string;
    capacidadPasajeros: number;
    autonomia:          number;
    potencia:           number;
    estado:             string;
    numPuertas:         number;
    precioBase:         number;
}